tested run of whole pipeline

+++ code/runApp.sh
@@ -1,23 +1,18 @@
 #!/bin/sh
 
 code=`dirname $0`
-
 usage="$0  <sampleID> <r1.fq> <r2.fq> <hg19.fasta> <manifest.txt>"
-
 if [ $# -lt 5 ];then
 	echo $usage
 	echo "Code repo is $code"
 	exit
 fi
 
-
 if [ ! -d $code ];then
 	echo No code repo found for scripts
 	echo exiting
 	exit
 fi
-
-
 #inputs
 #sample ID
 sid="$1"
@@ -32,7 +27,7 @@ outdir="$sid.outdir"
 variantdir="$outdir/variants"
 
 #created by App
-GENOMEIDX="hg19.fasta.nix"
+GENOMEIDX="$outdir/genome.nix"
 ampliconbed="$outdir/manifest2bed/ampliconNovoalign.bed"
 amplicon_noprimer="$outdir/manifest2bed/ampliconNoPrimerRegion.bed"
 variants="$outdir/variants.vcf.gz"
@@ -40,27 +35,47 @@ bam="$outdir/alignments.bam"
 
 mkdir -p $outdir $variantdir
 
-echo `date` Indexing the genome
-novoindex $GENOMEIDX $REFGENOME 
+if [ ! -e $GENOMEIDX ]; then
+	echo `date` Indexing the genome
+	novoindex $GENOMEIDX $REFGENOME 
+fi
 
+if [ ! -e $ampliconbed ];then
 #ManifestConversion
 perl $code/manifestToAmplicon.pl $manifest -outdir $outdir/manifest2bed/
-
+fi
 
 #Alignment
-echo `date` Align and Sorting
-novoalign -d $GENOMEIDX -f $fastqs --amplicons $ampliconbed -oSAM  2>novoalign.std_err.log |  samtools view -uS - > $outdir/novoalign.raw.sam
-novosort $outdir/novoalign.raw.sam -o $bam -i $outdir/novoalign.raw.sam  2>log.txt 
-rm  $outdir/novoalign.raw.sam
+if [ ! -e $bam ];then
+	echo `date` Align and Sorting
+	novoalign -d $GENOMEIDX -f $fastqs --amplicons $ampliconbed -oSAM -r Random -k  2> $outdir/novoalign_log.txt |  samtools view -uS - > $outdir/novoalign.raw.sam
+	novosort $outdir/novoalign.raw.sam -o $bam -i $outdir/novoalign.raw.sam  2>log.txt 
+	rm  $outdir/novoalign.raw.sam
+fi
 
 #Coverage Analysis
-intersectBed -abam $bam  -b  $amplicon_noprimer  -f 0.99999 -bed -wa -wb |sort -k 10,10 | groupBy -g 10 -c 1 -o count > $outdir/novo_coverage.tsv
+
+if [ ! -e "$outdir/novo_coverage.tsv" ];then
+echo `date` Calculating Coverage
+namecol=16
+intersectBed -abam $bam  -b  $amplicon_noprimer  -f 0.99999 -bed -wa -wb |sort -k $namecol,$namecol | groupBy -g $namecol -c 1 -o count > $outdir/novo_coverage.tsv
 #Plot graph
 #Rscript coverage.rscript
+fi
 
 #Variants calling
 mkdir -p $variantdir
-perl $code/callAmpliconVariants.pl -percent 0.99999 -cov 50 $amplicon_noprimer -bam $outdir/novoalign.bam -genome $REFGENOME -out $variantdir/ | parallel -j+0
+perl $code/callAmpliconVariants.pl \
+ -percent 0.99999 \
+ -cov 50 \
+  $amplicon_noprimer \
+ -bam $bam  \
+ -genome $REFGENOME \
+ -out $variantdir/ | parallel -j+0
+#index files
+ls $variantdir/*.gz | parallel -j+0 tabix -p vcf  {}
+
+#merge and index with Tabix
 bcftools merge --force-samples -o $variants -Oz $variantdir/*gz
 tabix -p vcf $variants
